requirements/base.txt
alembic
dotenv
fastapi
uvicorn
fastapi-users[sqlalchemy]
asyncpg
psycopg2-binary
fastapi-cache2[redis]